// src/user/domain/entity/address/country/enum/country-name.enum.ts
enum CountryNameEnum {
  ARGENTINA = 'Argentina',
  BOLIVIA = 'Bolivia',
  BRAZIL = 'Brazil',
  CHILE = 'Chile',
  COLOMBIA = 'Colombia',
  ECUADOR = 'Ecuador',
  GUYANA = 'Guyana',
  PARAGUAY = 'Paraguay',
  PERU = 'Peru',
  SURINAME = 'Suriname',
  URUGUAY = 'Uruguay',
  VENEZUELA = 'Venezuela',
}

export default CountryNameEnum;